+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Add `gcc_linux-64` and `gxx_linux-64` to R grading image dependencies per [#819](https://github.com/ucbds-infra/otter-grader/issues/819)
 * Fixed a bug where the loop variable used in Otter's generated grading code overwrote variables defined by the notebook per [#817](https://github.com/ucbds-infra/otter-grader/issues/817)
 * Add `--pickle-results` option to Otter Run to output the results pickle file per [#818](https://github.com/ucbds-infra/otter-grader/issues/818)
+* Added `setuptools` to Otter's required dependencies per [#823](https://github.com/ucbds-infra/otter-grader/issues/823)
 
 **v5.5.0:**