+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 **v5.6.0 (unreleased):**
 
-* Updated Otter Grade CSV to indicate which notebooks timeout [#813](https://github.com/ucbds-infra/otter-grader/issues/813)
+* Updated Otter Grade CSV to indicate which notebooks timeout per [#813](https://github.com/ucbds-infra/otter-grader/issues/813)
 * Updated Otter Grade CSV to include the number of points per question in the first row
 * Updated Otter Grade CSV to include total points column
 * Updated Otter Grade CSV to round percentages to four decimal places